Add an Authentication Service
This introduces an Authenticate function which will verify that user
credentials are correct. If they are, the service returns a json web token
which contains the id, firstname, lastname, and type of user (doctor, patient,
or nurse). The token should be sent to the backend on every subsequent request
to identify who is making a given request. The next step is creating a
Authorize function to validate tokens

Co-authored-by: Yann Cedric <ngadeuyann@yahoo.fr>